Added frontend for inventory

+++ Assets/Scripts/Attacks/BackWave.cs
@@ -22,7 +22,7 @@ public class BackWave : MonoBehaviour
     {
         _pos = transform.position;
         _range = GlobalScript.Gnox.RankRange < 4 ? 3 : 5;
-        _knock = 3 + Mathf.Floor(GlobalScript.Gnox.RankRange / 2);
+        _knock = 3 + Mathf.Floor((float)GlobalScript.Gnox.RankRange / 2);
         //get sphere of rigidbodys around the player
         _objects = Physics2D.OverlapCircleAll(GlobalScript.Gnox.Rigid.position, _range).ToList();
